pyproject.toml
[project]
name = "agent-workshop"
version = "0.1.0"
description = ""
requires-python = ">=3.12"
dependencies = [
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.8",
    "mysql-connector-python>=9.0.0",
    "pydantic-settings>=2.13.1",
    "streamlit>=1.54.0",
    "langchain>=1.0.0",
    "fastapi>=0.135.1",
    "uvicorn>=0.41.0",
]
[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests that need external services (e.g. MySQL/KVKart)",
]

[tool.pylint.messages_control]
disable = ["redefined-outer-name", "unused-argument"]
